% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.R
\name{AFM}
\alias{AFM}
\title{A-F-M}
\usage{
AFM(
  A,
  F,
  M,
  ternary = TRUE,
  twostage = TRUE,
  plot = TRUE,
  kde = TRUE,
  decision = TRUE,
  bty = "n",
  asp = 1,
  xpd = FALSE,
  bw = "nrd0",
  dlty = 2,
  dlwd = 1.5,
  dcol = "blue",
  padding = 0.15,
  ...
)
}
\arguments{
\item{A}{vector with (Na\eqn{_2}O+K\eqn{_2}O) concentrations (in
wt\%)}

\item{F}{vector with (FeO + FeO\eqn{_2}O\eqn{_3}) concentrations
(in wt\%)}

\item{M}{vector with MgO concentrations (in wt\%)}

\item{ternary}{logical. If \code{FALSE}, produces a logratio plot.}

\item{twostage}{logical. If \code{TRUE}, applies the two-stage
magma evolution model of Vermeesch and Pease (2021). Otherwise
applies the single stage model.}

\item{plot}{logical. If \code{FALSE}, omits the graphical output
and only returns the Bowen-Fenner indices of the sample.}

\item{kde}{logical. If \code{TRUE}, adds a kernel density estimate
of the Bowen-Fenner indices on a radial scale. Not used if
\code{ternary=FALSE}.}

\item{decision}{logical. If \code{TRUE}, adds the decision boundary
between the tholeiitic and calc-alkaline magma series to the
plot.}

\item{bty}{A character string that determines the type of `box'
which is drawn about plots. See \code{par}.}

\item{asp}{the y/x aspect ratio, see \code{plot.window}.}

\item{xpd}{clip the plot region? See \code{par} for further details.}

\item{bw}{the smoothing bandwidth to be used for the kernel density
estimate. See \code{density}.}

\item{dlty}{line type of the decision boundary.}

\item{dlwd}{line width of the decision boundary.}

\item{dcol}{colour of the decision boundary.}

\item{padding}{fractional measure of distance between the data and
the kernel density estimate.}

\item{...}{additional arguments for the generic \code{points}
function.}
}
\value{
the Bowen-Fenner indices of the samples, where positive
    values indicate calc-alkaline and negative numbers tholeiitic
    compositions (Vermeesch and Pease, 2021).
}
\description{
A-F-M diagram
}
\examples{
data(cath,package='GeoplotR')
oldpar <- par(mfrow=c(2,2))
on.exit(par(oldpar))

A <- cath[,'Na2O']+cath[,'K2O']
F <- cath[,'FeOT']
M <- cath[,'MgO']
tern <- c(TRUE,FALSE,TRUE,FALSE)
stag <- c(TRUE,TRUE,FALSE,FALSE)
for (i in 1:4){
    AFM(A,F,M,ternary=tern[i],decision=TRUE,twostage=stag[i],
        pch=21,bw=0.2,padding=0.15,bg=cath[,'affinity'])
}

dev.new()
CA <- (cath$affinity=='ca')
TH <- (cath$affinity=='th')
fitCA <- AFM(A[CA],F[CA],M[CA],plot=FALSE)
fitTH <- AFM(A[TH],F[TH],M[TH],plot=FALSE)
d1 <- density(fitCA)
d2 <- density(fitTH)
matplot(cbind(d1$x,d2$x),cbind(d1$y,d2$y),type='l',lty=1)
legend('topright',legend=c('Cascades','Iceland'),lty=1)

}
\references{
Irvine, T. N. and Baragar, W. A guide to the chemical
    classification of the common volcanic rocks. Canadian Journal
    of Earth Sciences, 8(5):523--548, 1971.

Kuno, H. Differentiation of basalt magmas. Basalts: The Poldervaart
    treatise on rocks of basaltic composition, pages 623--688, 1968.

Vermeesch, P. and Pease, V. A genetic classification of the
tholeiitic and calc-alkaline magma series, Geochemical Perspective
Letters (in review).
}
