% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz.R
\name{AnAbOr}
\alias{AnAbOr}
\title{An-Ab-Or}
\usage{
AnAbOr(An = NULL, Ab = NULL, Or = NULL, show.labels = TRUE, ...)
}
\arguments{
\item{An}{vector with normative anorthite concentrations (wt\%)}

\item{Ab}{vector with normative albite concentrations (wt\%)}

\item{Or}{vector with normative orthoclase concentrations (wt\%)}

\item{show.labels}{logical. If \code{TRUE}, labels the
discrimination fields on the plot.}

\item{...}{additional arguments for the generic \code{points}
function.}
}
\value{
a list with labels, indicating whether the input data are
    tonalite, trondhjemite, granodiorite, granite.
}
\description{
An-Ab-Or diagram for tonalite, trondhjemite,
    granodiorite, and granite according to Barker (1979).
}
\examples{
AnAbOr(An=c(70,75,73),Ab=c(20,10,27),Or=c(10,15,0))
}
\references{
Barker, F., 1979, Trondhjemite: Definition,
    environment, and hypotheses of origin: p. 1-12, in Barker, F.,
    ed., Trondhjemites, Dacites, and Related Rocks, Elsevier,
    Amsterdam, 659 p.
}
