% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATM.R
\name{ATM}
\alias{ATM}
\title{A-T-M}
\usage{
ATM(
  A,
  T,
  M,
  ternary = TRUE,
  plot = TRUE,
  kde = TRUE,
  decision = TRUE,
  bty = "n",
  asp = 1,
  xpd = FALSE,
  bw = "nrd0",
  dlty = 2,
  dlwd = 1.5,
  dcol = "blue",
  padding = 0.15,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{A}{vector with (Na\eqn{_2}O+K\eqn{_2}O) concentrations (in
wt\%)}

\item{T}{vector with TiO\eqn{_2} concentrations}

\item{M}{vector with MgO concentrations (in wt\%)}

\item{ternary}{logical. If \code{FALSE}, produces a logratio plot.}

\item{plot}{logical. If \code{FALSE}, omits the graphical output
and only returns the Bowen-Fenner indices of the sample.}

\item{kde}{logical. If \code{TRUE}, adds a kernel density estimate
of the Bowen-Fenner indices on a radial scale. Not used if
\code{ternary=FALSE}.}

\item{decision}{logical. If \code{TRUE}, adds the decision boundary
between the tholeiitic and calc-alkaline magma series to the
plot.}

\item{bty}{A character string that determines the type of `box'
which is drawn about plots. See \code{par}.}

\item{asp}{the y/x aspect ratio, see \code{plot.window}.}

\item{xpd}{clip the plot region? See \code{par} for further
details.}

\item{bw}{the smoothing bandwidth to be used for the kernel density
estimate. See \code{density}.}

\item{dlty}{line type of the decision boundary.}

\item{dlwd}{line width of the decision boundary.}

\item{dcol}{colour of the decision boundary.}

\item{padding}{fractional measure of distance between the data and
the kernel density estimate.}

\item{xlim}{the x axis limits of the plot (see \code{plot.default}
for details).}

\item{ylim}{the y axis limits of the plot (see \code{plot.default}
for details).}

\item{...}{additional arguments for the generic \code{points}
function.}
}
\value{
the Bowen-Fenner indices of the samples, where positive
    values indicate calc-alkaline and negative numbers tholeiitic
    compositions (Vermeesch and Pease, 2021).
}
\description{
A-T-M diagram
}
\examples{
data(cath,package='GeoplotR')

oldpar <- par(mfrow=c(2,2),mar=c(4,4,0,0))
on.exit(par(oldpar))

A <- cath[,'Na2O']+cath[,'K2O']
T <- cath[,'TiO2']
M <- cath[,'MgO']

ATM(A,T,M,decision=FALSE,pch=21,bg='white',cex=0.5)

ATM(A,T,M,ternary=TRUE,pch=21,bw=0.2,
    padding=0.15,bg=cath[,'affinity'])

ATM(A,T,M,ternary=FALSE,decision=TRUE,kde=TRUE,
    pch=21,bw=0.2,padding=0.15,bg=cath[,'affinity'])
   
CA <- (cath$affinity=='ca')
TH <- (cath$affinity=='th')
fitCA <- ATM(A[CA],T[CA],M[CA],plot=FALSE)
fitTH <- ATM(A[TH],T[TH],M[TH],plot=FALSE)
d1 <- density(fitCA)
d2 <- density(fitTH)
matplot(cbind(d1$x,d2$x),cbind(d1$y,d2$y),
        type='l',lty=1,xlab='BF',ylab='')
legend('topleft',legend=c('Cascades','Iceland'),lty=1)

}
\references{
Vermeesch, P. and Pease, V. A genetic classification of the
tholeiitic and calc-alkaline magma series, Geochemical Perspective
Letters.
}
