% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BF.R
\name{BF}
\alias{BF}
\title{Bowen-Fenner index}
\usage{
BF(A, F, M, T, twostage = TRUE)
}
\arguments{
\item{A}{vector with (Na\eqn{_2}O+K\eqn{_2}O) concentrations (in
wt\%)}

\item{F}{vector with (FeO + FeO\eqn{_2}O\eqn{_3}) concentrations
(in wt\%)}

\item{M}{vector with MgO concentrations (in wt\%)}

\item{T}{vector with TiO\eqn{_2} concentrations (in wt\%)}

\item{twostage}{logical. If \code{TRUE}, uses the two-stage magma
evolution model. If \code{TRUE}, uses a simple one-stage
model. Only used if \code{T} has been omitted.}
}
\value{
the Bowen-Fenner indices of the samples, where positive
    values indicate calc-alkaline and negative numbers tholeiitic
    compositions.

If \code{F} is provided, \code{T} is omitted, and \code{twostage}
is \code{FALSE}: returns the \eqn{BF_1} index of Vermeesch and
Pease (2021).

If \code{F} is provided, \code{T} is omitted, and \code{twostage}
is \code{TRUE}: returns the \eqn{BF_2^F} index.

If \code{F} is omitted and \code{T} is provided, returns the
\eqn{BF_2^T} index.

If both \code{F} and \code{T} are provided, returns the \eqn{BF_2}
index, which is the mean of \eqn{BF_2^F} and \eqn{BF_2^T}.
}
\description{
Bowen-Fenner index for tholeiitic and calc-alkaline
    magma series
}
\examples{
data(cath,package='GeoplotR')

oldpar <- par(mfrow=c(2,2),mar=c(4,4.5,2,0))
on.exit(par(oldpar))

A <- cath[,'Na2O']+cath[,'K2O']
F <- cath[,'FeOT']
T <- cath[,'TiO2']
M <- cath[,'MgO']

CA <- (cath$affinity=='ca')
TH <- (cath$affinity=='th')

dCA <- list()
dTH <- list()

titles <- c('BF1','BF2F','BF2T','BF2')

dCA[[1]] <- density(BF(A=A[CA],F=F[CA],M=M[CA],twostage=FALSE))
dTH[[1]] <- density(BF(A=A[TH],F=F[TH],M=M[TH],twostage=FALSE))

dCA[[2]] <- density(BF(A=A[CA],F=F[CA],M=M[CA],twostage=TRUE))
dTH[[2]] <- density(BF(A=A[TH],F=F[TH],M=M[TH],twostage=TRUE))

dCA[[3]] <- density(BF(A=A[CA],M=M[CA],T=T[CA],twostage=TRUE))
dTH[[3]] <- density(BF(A=A[TH],M=M[TH],T=T[TH],twostage=TRUE))

dCA[[4]] <- density(BF(A=A[CA],F=F[CA],M=M[CA],T=T[CA],twostage=TRUE))
dTH[[4]] <- density(BF(A=A[TH],F=F[TH],M=M[TH],T=T[TH],twostage=TRUE))

for (i in 1:4){
    matplot(cbind(dCA[[i]]$x,dTH[[i]]$x),
            cbind(dCA[[i]]$y,dTH[[i]]$y),
            type='l',lty=1,col=c('red','black'),
            xlab='BF',ylab='')
    title(titles[i])
    legend('topleft',legend=c('Cascades','Iceland'),
           lty=1,col=c('red','black'))
}

}
\references{
Vermeesch, P. and Pease, V. 2021, A genetic
    classification of the tholeiitic and calc-alkaline magma
    series, Geochemical Perspective Letters.
}
