% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{cart}
\alias{cart}
\title{tectonic discrimination by decision tree}
\usage{
cart(
  dat,
  option = 1,
  type = c("class", "prob", "vector", "matrix"),
  plot = FALSE,
  ...
)
}
\arguments{
\item{dat}{a data frame or matrix with the following columns
    (missing data are allowed):

If \code{option=1}: \code{SiO2} (as wt\%) or \code{Si} (as ppm),
\code{TiO2} (as wt\%) or \code{Ti} (as ppm), \code{CaO} (as wt\%)
or \code{Ca} (as ppm), \code{Fe2O3} (as wt\%) or \code{Fe3} (as
ppm), \code{MgO} (as wt\%) or \code{Mg} (as ppm), \code{K2O} (as
wt\%) or \code{K} as ppm, \code{La}, \code{Pr}, \code{Nd},
\code{Sm}, \code{Gd}, \code{Tb}, \code{Yb}, \code{Lu}, \code{V},
\code{Ni}, \code{Rb}, \code{Sr}, \code{Y}, \code{Hf}, \code{Th}
(all as ppm), \code{Sr87/Sr86} and \code{Pb206/Pb204}.

If \code{option=2}: \code{TiO2} (as wt\%) or \code{Ti} (as ppm),
\code{La}, \code{Ce}, \code{Pr}, \code{Nd}, \code{Sm}, \code{Gd},
\code{Tb}, \code{Dy}, \code{Ho}, \code{Er}, \code{Tm}, \code{Yb},
\code{Lu}, \code{Sc}, \code{Y}, \code{Zr}, \code{Nb}, \code{Hf},
\code{Ta}, \code{Pb}, \code{Th}, \code{U} (all as ppm),
{Nd143/Nd144}, \code{Sr87/Sr86}, \code{Pb206/Pb204},
\code{Pb207/Pb204} and \code{Pb208/Pb204}.

If \code{option=3}: \code{TiO2} (as wt\%) or \code{Ti} (as ppm),
\code{La}, \code{Sm}, \code{Nd}, \code{Gd}, \code{Yb}, \code{Sc},
\code{V}, \code{Sr}, {Y}, \code{Zr}, \code{Nb}, \code{Th}, \code{U}
(all as ppm).}

\item{option}{numerical. If \code{option=1}, uses all major and
trace element concentrations and isotopic ratios, if
\code{option=2}, uses high field strength element
concentrations and isotopic ratios, if \code{option=3}, uses
high field strength element ratios.}

\item{type}{character string denoting the type of predicted value
returned.}

\item{plot}{logical. If \code{TRUE}, produces a dendrogram with the
leaves marking the number of classified samples.}

\item{...}{optional arguments to \code{text.rpart}, except for
\code{xpd} and \code{use.n}, which have been set to \code{NA}
and \code{TRUE}, respectively and cannot be changed.}
}
\value{
If \code{type="class"}: a factor of classifications based
    on the responses.

If \code{type="prob"}: a matrix of class probabilities.

If \code{type="vector"}: vector of predicted responses.

If \code{type="matrix"}: a matrix of the full responses, i.e. the
concatenation of at least the predicted class, the class counts at
that node in the fitted tree, and the class probabilities.
}
\description{
Classification And Regression Trees (CART) of basalts
    from ocean islands (OIB), mid-oceanic ridges (MORB) and island
    arcs (IAB).
}
