% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{Catanorm}
\alias{Catanorm}
\title{Barth-Niggli's molecular norm}
\usage{
Catanorm(WR)
}
\arguments{
\item{WR}{a data frame with whole rock data}
}
\value{
a data frame with the inferred normative mineralogical composition.
}
\description{
Calculates the cationic norm of Niggli (1948)
}
\examples{
data(test, package='GeoplotR')
norm  <-  Catanorm(test)
}
\references{
Janousek, V., Farrow, C.M. and Erban, V.,
2006. Interpretation of whole-rock geochemical data in igneous
geochemistry: introducing Geochemical Data Toolkit
(GCDkit). Journal of Petrology, 47(6), pp.1255-1259.

Niggli P (1948) Gesteine und Minerallagerstatten. Birkhauser, Basel, p. 1-540
}
\author{
Vojtech Janousek
}
