% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{cath}
\alias{cath}
\title{A-F-M data}
\description{
Calc-alkaline and tholeiitic data from the Cascade
    Mountains and Iceland
}
\examples{
data(cath,package='GeoplotR')
ca <- cath[cath$affinity=='ca',]
A <- cath[,'Na2O']+cath[,'K2O']
F <- cath[,'FeOT']
M <- cath[,'MgO']
AFM(A,F,M,pch=16,col=cath$affinity)
legend('topleft',legend=c('Cascades','Iceland'),pch=16,col=c(1,2))
}
\references{
Vermeesch, P. and Pease, V. A genetic classification of
    the tholeiitic and calc-alkaline magma series, Geochemical
    Perspective Letters (in review).
}
\keyword{data}
