% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{CIPW}
\alias{CIPW}
\title{CIPW norm}
\usage{
CIPW(
  wrdata,
  normsum = FALSE,
  cancrinite = FALSE,
  spinel = FALSE,
  complete.results = FALSE
)
}
\arguments{
\item{wrdata}{a data frame with whole rock data}

\item{normsum}{logical. Recast the norm to 100\%?}

\item{cancrinite}{logical. Is cancrinite present?}

\item{spinel}{logical. Shall be spinel calculated for rocks with SiO\code{_2 < 45}\%?}

\item{complete.results}{logical. Remove rows with incomplete results?}
}
\value{
a data frame with the inferred normative mineralogical composition.
}
\description{
Cross-Iddings-Pirsson-Washington norm
}
\examples{
data(test, package='GeoplotR')
norm <- CIPW(test)
}
\references{
Janousek, V., Farrow, C.M. and Erban, V.,
2006. Interpretation of whole-rock geochemical data in igneous
geochemistry: introducing Geochemical Data Toolkit
(GCDkit). Journal of Petrology, 47(6), pp.1255-1259.

Hutchison, C.S., 1974. Laboratory handbook of petrographic
techniques (Vol. 527). New York: Wiley.
}
\author{
Vojtech Janousek
}
