% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{conconv}
\alias{conconv}
\alias{wtpct2ppm}
\alias{ppm2wtpct}
\title{concentration unit conversion}
\usage{
conconv(x, oxide, wtpct2ppm = TRUE)

wtpct2ppm(x, oxide)

ppm2wtpct(x, oxide)
}
\arguments{
\item{x}{scalar or vector of wt\% or ppm values}

\item{oxide}{Oxide of the element; one of \code{'SiO2'},
\code{'TiO2'}, \code{'Al2O3'}, \code{'Fe2O3'}, \code{'FeO'},
\code{'CaO'}, \code{'MgO'}, \code{'MnO'}, \code{'K2O'},
\code{'Na2O'}, \code{'P2O5'}, \code{'Y2O3'}, \code{'ZrO2'}}

\item{wtpct2ppm}{logical. If \code{TRUE}, converts wt\% to ppm,
otherwise converts ppm to wt\%.}
}
\value{
converted values
}
\description{
Converts from percentage oxide by weight to parts per
    million of the element and vice versa. \code{wtpct2ppm} and
    \code{ppm2wtpct} are special cases.
}
\examples{
ppm <- wtpct2ppm(0.308,'TiO2')
wtpct <- ppm2wtpct(ppm,'TiO2')
ppm <- wtpct2ppm(0.308,'TiO2')
wtpct <- ppm2wtpct(ppm,'TiO2')
}
