% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{CrY}
\alias{CrY}
\title{Cr-Y}
\usage{
CrY(Cr, Y, xlim = c(1, 100), ylim = c(1, 10000), show.labels = TRUE, ...)
}
\arguments{
\item{Cr}{vector with Cr concentrations (ppm)}

\item{Y}{vector with Y concentrations (ppm)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{show.labels}{logical. If \code{TRUE}, labels the polygonal
decision fields with text strings.}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}
}
\value{
a vector with tectonic affinities
}
\description{
Cr vs Y discriminant for MORB, IAT, and boninites of
    Dilek et al. (2007)
}
\examples{
data(test,package='GeoplotR')
CrY(Cr=test[,'Cr'],Y=test[,'Y'])
}
\references{
Dilek, Y., Furnes, H., and Shallo, M., 2007,
    Suprasubduction zone ophiolite formation along the periphery of
    Mesozoic Gondwana. Gondwana Research, v. 11, p. 453-475.
}
