% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hollocher2012.R
\name{Hollocher2012}
\alias{Hollocher2012}
\alias{NbLaYb}
\alias{ThNbLaYb}
\title{Hollocher et al. (2012)}
\usage{
NbLaYb(
  Nb = NULL,
  La = NULL,
  Yb = NULL,
  xlim = NULL,
  ylim = NULL,
  show.labels = TRUE,
  short = FALSE,
  ...
)

ThNbLaYb(
  Th = NULL,
  Nb = NULL,
  La = NULL,
  Yb = NULL,
  xlim = NULL,
  ylim = NULL,
  show.labels = TRUE,
  short = FALSE,
  ...
)
}
\arguments{
\item{Nb}{vector with Nb concentrations (ppm)}

\item{La}{vector with La concentrations (ppm)}

\item{Yb}{vector with Yb concentrations (ppm)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{show.labels}{logical. If \code{TRUE}, labels the polygonal
fields with their respective tectonic affinities.}

\item{short}{logical. If \code{TRUE}, uses abbreviated labels.}

\item{...}{additional arguments to the generic \code{points}
function.}

\item{Th}{vector with Th concentrations (ppm)}
}
\value{
a vector with tectonic affinities
}
\description{
Discriminants for MORB, ocean island, and a variety of
    arc-type basalts
}
\examples{
data(test,package='GeoplotR')
NbLaYb(Nb=test[,'Nb'],La=test[,'La'],Yb=test[,'Yb'])
ThNbLaYb(Th=test[,'Th'],Nb=test[,'Nb'],La=test[,'La'],Yb=test[,'Yb'])
}
\references{
Hollocher et al., 2012 Discriminants for MORB, ocean
    island, and a variety of arc-type basalts
}
