% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{LaYb}
\alias{LaYb}
\title{La/Yb-Yb}
\usage{
LaYb(La_n, Yb_n, xlim = c(0, 25), ylim = c(0, 160), show.labels = TRUE, ...)
}
\arguments{
\item{La_n}{vector with chondritic normalised La concentrations}

\item{Yb_n}{vector with chondritic normalised Yb concentrations}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{show.labels}{logical. If \code{TRUE}, labels the polygonal
decision fields with text strings.}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}
}
\value{
a vector with tectonic affinities
}
\description{
(La/Yb)_N vs. Yb_N diagram of Martin (1986), which
    discriminates between Archean TTG suite and more modern
    adakites on the one hand, and classical island arcs on the
    other hand.
}
\examples{
data(test,package='GeoplotR')
LaYb(La_n=100,Yb_n=10)
}
\references{
Martin, H., 1986. Effect of steeper Archean geothermal
    gradient on geochemistry of subduction-zone magmas. Geology,
    14(9), pp.753-756.
}
