% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vermeesch2006.R
\name{NbZrY}
\alias{NbZrY}
\title{Nb-Zr-Y}
\usage{
NbZrY(
  Nb = NULL,
  Zr = NULL,
  Y = NULL,
  type = c("Meschede", "LDA", "QDA"),
  ternary = TRUE,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  ...
)
}
\arguments{
\item{Nb}{vector with Nb concentrations (ppm)}

\item{Zr}{vector with Zr concentrations (ppm)}

\item{Y}{vector with Y concentrations (ppm)}

\item{type}{either \code{'LDA'} for linear discriminant analysis,
\code{'QDA'} for quadratic discriminant analysis, or
\code{'Pearce'} for the nominal decision boundaries of Meschede
(1986).}

\item{ternary}{logical. If \code{FALSE}, produces a logratio plot.
Only used if \code{type} is \code{LDA} or \code{QDA}.}

\item{pch}{plot character. See \code{?par} for details.}

\item{bg}{fill colour for the plot character.}

\item{show.labels}{logical. If \code{TRUE}, labels the
discrimination fields.}

\item{short}{logical. If \code{TRUE}, uses abbreviated labels for
the discrimination fields.}

\item{...}{additional arguments for the generic \code{points}
function.}
}
\value{
if \code{type='LDA'} or \code{type='QDA'}, a list with
    components \code{class}, \code{posterior} and \code{x};
    otherwise a table with tectonic affinities.
}
\description{
Nb-Zr-Y tectonic discrimination diagram
}
\examples{
data(test,package='GeoplotR')
NbZrY(Nb=test[,'Nb'],Zr=test[,'Zr'],Y=test[,'Y'])
}
\references{
Meschede, M., 1986. A method of discriminating between
    different types of mid-ocean ridge basalts and continental
    tholeiites with the Nb-Zr-Y diagram. Chemical geology, 56(3-4),
    pp.207-218.
}
