% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pearce1984.R
\name{Pearce1984}
\alias{Pearce1984}
\alias{YNb}
\alias{YNbRb}
\alias{YbTa}
\alias{YbTaRb}
\title{Pearce et al. (1984)}
\usage{
YNb(Y = NULL, Nb = NULL, show.labels = TRUE, xlim = NULL, ylim = NULL, ...)

YNbRb(
  Y = NULL,
  Nb = NULL,
  Rb = NULL,
  show.labels = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)

YbTa(Yb = NULL, Ta = NULL, xlim = NULL, ylim = NULL, show.labels = TRUE, ...)

YbTaRb(
  Yb = NULL,
  Ta = NULL,
  Rb = NULL,
  show.labels = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{Y}{vector with Y concentrations (ppm)}

\item{Nb}{vector with Nb concentrations (ppm)}

\item{show.labels}{logical. If \code{TRUE}, labels the polygonal
fields with their respective tectonic affinities.}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{...}{additional arguments to the generic \code{points}
function.}

\item{Rb}{vector with Rb concentrations (ppm)}

\item{Yb}{vector with Yb concentrations (ppm)}

\item{Ta}{vector with Ta concentrations (ppm)}
}
\value{
a vector with tectonic affinities
}
\description{
tectonic discrimination diagram for granites
}
\examples{
data(test,package='GeoplotR')
YNb(Y=test[,'Y'],Nb=test[,'Nb'])
YNbRb(Y=test[,'Y'],Nb=test[,'Nb'],Rb=test[,'Rb'])
YbTa(Yb=test[,'Yb'],Ta=test[,'Ta'])
YbTaRb(Yb=test[,'Yb'],Ta=test[,'Ta'],Rb=test[,'Rb'])
}
\references{
Pearce, J. A., Harris, N. B. W., and Tindle, A. G.,
    1984,
    "Trace Element Discrimination Diagrams for the Tectonic Interpretation of Granitic Rocks":
    Journal of Petrology, v. 25, no. 4, p. 956-983.
}
