% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz.R
\name{QAP}
\alias{QAP}
\title{QAP diagram}
\usage{
QAP(
  Q = NULL,
  A = NULL,
  P = NULL,
  pch = 21,
  bg = NULL,
  show.labels = TRUE,
  short = TRUE,
  ...
)
}
\arguments{
\item{Q}{vector with quartz concentrations (\%)}

\item{A}{vector with alkali feldspar concentrations (\%)}

\item{P}{vector with plagioclase feldspar concentrations (\%)}

\item{show.labels}{logical. If \code{TRUE}, labels the
discrimination fields on the plot.}

\item{short}{use short labels when using the additional argument
\code{show.labels=TRUE}.}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}
}
\value{
a vector with rock types
}
\description{
Streckeisen (1967)'s Quartz - Alkali feldspar -
    Plagioclase feldspar diagram for classification of felsic and
    intermediate plutonic rocks.
}
\examples{
QAP(Q=c(40,30),A=c(20,40),P=c(40,30))
}
\references{
Streckeisen, A., 1967, Classification and nomenclature
    of igneous rocks.  Nues Jarbuch fur Mineralogie Abhandlungen,
    v. 107, p. 144-240.
}
