% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAPF.R
\name{QAPF}
\alias{QAPF}
\alias{QAP}
\alias{FAP}
\title{QAPF diagram}
\usage{
QAPF(
  Q = NULL,
  A = NULL,
  P = NULL,
  F = NULL,
  pch = 21,
  bg = NULL,
  show.labels = TRUE,
  short = TRUE,
  ...
)

QAP(
  Q = NULL,
  A = NULL,
  P = NULL,
  pch = 21,
  bg = NULL,
  show.labels = TRUE,
  short = TRUE,
  ...
)

FAP(
  F = NULL,
  A = NULL,
  P = NULL,
  pch = 21,
  bg = NULL,
  show.labels = TRUE,
  short = TRUE,
  ...
)
}
\arguments{
\item{Q}{quartz (\%)}

\item{A}{alkali feldspar (\%)}

\item{P}{plagioclase feldspar (\%)}

\item{F}{feldspathoid (\%)}

\item{pch}{plot character. See \code{?par} for details.}

\item{bg}{fill colour of the plot symbols.}

\item{show.labels}{logical. If \code{TRUE}, labels the
discrimination fields on the plot.}

\item{short}{use short labels when using the additional argument
\code{show.labels=TRUE}.}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}
}
\value{
a vector with rock names
}
\description{
Streckeisen (1967)'s Quartz - Alkali feldspar -
    Plagioclase feldspar diagram for classification of felsic and
    intermediate plutonic rocks.
}
\examples{
   out <- QAPF(Q=c(50,40,0,0,80,60),
               A=c(10,20,50,0,0,20),
               P=c(40,40,50,20,20,20),
               F=c(0,0,0,80,0,0))
}
\references{
Streckeisen, A., 1967, Classification and nomenclature
    of igneous rocks.  Nues Jarbuch fur Mineralogie Abhandlungen,
    v. 107, p. 144-240.
}
