% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{SrY}
\alias{SrY}
\title{Sr/Y-Y}
\usage{
SrY(Sr = NULL, Y = NULL, xlim = NULL, ylim = NULL, show.labels = TRUE, ...)
}
\arguments{
\item{Sr}{vector with Sr concentrations (in ppm)}

\item{Y}{vector with Y concentrations (in ppm)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{show.labels}{logical. If \code{TRUE}, labels the polygonal
decision fields with text strings.}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}
}
\value{
a vector with tectonic affinities
}
\description{
(Sr/Y) vs. Y diagram of Drummond (1990), which
    discriminates between Archean TTG suite and more modern
    adakites on the one hand, and classical island arcs on the
    other hand.
}
\examples{
data(test,package='GeoplotR')
SrY(Sr=1000,Y=10)
}
\references{
Drummond MS, Defant MJ (1990) A model for
    trondhjemite-tonalite-dacite genesis and crustal growth via
    slab melting: Archean to modern comparisons. J Geophys Res
    95:21503– 21521
}
