% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{TAS}
\alias{TAS}
\title{TAS diagram}
\usage{
TAS(Na2O, K2O, SiO2, plot = TRUE, labels = FALSE, volcanic = TRUE, ...)
}
\arguments{
\item{Na2O}{vector with \eqn{Na_2O} concentrations (wt\%)}

\item{K2O}{vector with \eqn{K_2O} concentrations (wt\%)}

\item{SiO2}{vector with \eqn{SiO_2} concentrations (wt\%)}

\item{plot}{logical. If \code{FALSE}, suppresses the graphical
output.}

\item{labels}{logical. Label to mark the different fields of the
TAS diagram.}

\item{volcanic}{logical. Set to \code{FALSE} for plutonic rocks.}

\item{...}{additional arguments to the generic \code{points}
function.}
}
\value{
a vector with rock types
}
\description{
Total Alkali Silica diagram
}
\examples{
data(test,package='GeoplotR')
TAS(Na2O=test[,'NA2O(WT\%)'],
    K2O=test[,'K2O(WT\%)'],
    SiO2=test[,'SIO2(WT\%)'])
}
