% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{test}
\alias{test}
\title{test data}
\description{
MOIB, OIB and IAB test set from Vermeesch (2006)
}
\examples{
data(test,package='GeoplotR')
TiZrY(test[,'TiO2'],test[,'Zr'],test[,'Y'],units=c('wt\%','ppm','ppm'))
}
\references{
Vermeesch, P., 2006, Tectonic discrimination diagrams
    revisited: Geochemistry, Geophysics, and Geosystems, 7, Q06017,
    doi:10.1029/2005GC001092
}
\keyword{data}
