% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{ThCo}
\alias{ThCo}
\title{Th-Co}
\usage{
ThCo(Th = NULL, Co = NULL, xlim = NULL, ylim = NULL, show.labels = TRUE, ...)
}
\arguments{
\item{Th}{vector with Th concentrations (ppm)}

\item{Co}{vector with Co concentrations (ppm)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{show.labels}{logical. If \code{TRUE}, labels the polygonal
decision fields with text strings.}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}
}
\value{
a vector with tectonic affinities
}
\description{
Th vs Co discrimination diagram of Hastie et
    al. (2007) for tholeiitic, calc-alkaline, and more alkaline arc
    series, and basaltic, andesitic, and rhyolitic compositions.
}
\examples{
data(test,package='GeoplotR')
ThCo(Th=test[,'Th'],Co=test[,'Co'])
}
\references{
Hastie, A.R., Kerr, A.C., Pearce, J.A., Mitchell, S.F.,
    2007, Classification of Altered Volcanic Island Arc Rocks using
    Immobile Trace Elements: Development of the Th-Co
    Discrimination Diagram.  Journal of Petrology, v. 48,
    p. 2341-2357.
}
