% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vermeesch2006.R
\name{ThTaHf}
\alias{ThTaHf}
\title{Th-Ta-Hf}
\usage{
ThTaHf(
  Th = NULL,
  Ta = NULL,
  Hf = NULL,
  type = c("Wood", "LDA", "QDA"),
  ternary = TRUE,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  ...
)
}
\arguments{
\item{Th}{vector with Th concentrations (ppm)}

\item{Ta}{vector with Ta concentrations (ppm)}

\item{Hf}{vector with Hf concentrations (ppm)}

\item{type}{either \code{'LDA'} for linear discriminant analysis,
\code{'QDA'} for quadratic discriminant analysis, or
\code{'Wood'} for the nominal decision boundaries of Wood
(1980).}

\item{ternary}{logical. If \code{FALSE}, produces a logratio plot.
Only used if \code{type} is \code{LDA} or \code{QDA}.}

\item{pch}{plot character. See \code{?par} for details.}

\item{bg}{fill colour for the plot character.}

\item{show.labels}{logical. If \code{TRUE}, labels the
discrimination fields.}

\item{short}{logical. If \code{TRUE}, uses abbreviated labels for
the discrimination fields.}

\item{...}{additional arguments for the generic \code{points}
function.}
}
\value{
if \code{type='LDA'} or \code{type='QDA'}, a list with
    components \code{class}, \code{posterior} and \code{x};
    otherwise a table with tectonic affinities.
}
\description{
Th-Ta-Hf tectonic discrimination diagram
}
\examples{
data(test,package='GeoplotR')
ThTaHf(Th=test[,'Th'],Ta=test[,'Ta'],Hf=test[,'Hf'])
}
\references{
Wood, D.A., 1980. The application of a Th-Hf-Ta diagram
    to problems of tectonomagmatic classification and to
    establishing the nature of crustal contamination of basaltic
    lavas of the British Tertiary Volcanic Province. Earth and
    Planetary Science Letters, 50(1), pp.11-30.
}
