% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vermeesch2006.R
\name{TiV}
\alias{TiV}
\title{Ti-V}
\usage{
TiV(
  Ti = NULL,
  V = NULL,
  type = c("LDA", "QDA", "Shervais"),
  ternary = FALSE,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  ...
)
}
\arguments{
\item{Ti}{vector with Ti concentrations (ppm)}

\item{V}{vector with V concentrations (ppm)}

\item{type}{either \code{'LDA'} for linear discriminant analysis,
\code{'QDA'} for quadratic discriminant analysis, or
\code{'Shervais'} for the nominal decision boundaries of
Shervais (1982).}

\item{ternary}{logical. If \code{TRUE}, produces a ternary diagram.
Only used if \code{type} is \code{LDA} or \code{QDA}.}

\item{pch}{plot character. See \code{?par} for details.}

\item{bg}{fill colour for the plot character.}

\item{show.labels}{logical. If \code{TRUE}, labels the
discrimination fields.}

\item{short}{logical. If \code{TRUE}, uses abbreviated labels for
the discrimination fields.}

\item{...}{additional arguments for the generic \code{points}
function.}
}
\value{
if \code{type='LDA'} or \code{type='QDA'}, a list with
    components \code{class}, \code{posterior} and \code{x};
    otherwise a table with labels for \code{MORB}, \code{IAB} and
    \code{OIB}.
}
\description{
Ti-V tectonic discrimination diagram
}
\examples{
data(test,package='GeoplotR')
TiV(Ti=wtpct2ppm(test[,'TiO2']),V=test[,'V'],type='Shervais')
}
\references{
Shervais, J.W., 1982. Ti-V plots and the petrogenesis
    of modern and ophiolitic lavas. Earth and Planetary Science
    Letters, 59(1), pp.101-118.
}
