% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{training}
\alias{training}
\title{training data}
\description{
MOIB, OIB and IAB training set from Vermeesch (2006)
}
\examples{
data(training,package='GeoplotR')
TiZrY(Ti=training[,'TIO2(WT\%)'],Zr=training[,'ZR(PPM)'],
      Y=training[,'Y(PPM)'],units=c(2,1,1))
}
\references{
Vermeesch, P., 2006, Tectonic discrimination diagrams
    revisited: Geochemistry, Geophysics, and Geosystems, 7, Q06017,
    doi:10.1029/2005GC001092
}
\keyword{data}
