% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vermeesch2006.R
\name{Vermeesch2006}
\alias{Vermeesch2006}
\alias{TiSiSr}
\alias{LuEuSr}
\alias{TiVSc}
\alias{NbNaSr}
\alias{TiSmV}
\alias{TiZrYSr}
\alias{Pearce1976}
\title{Vermeesch (2006)}
\usage{
TiSiSr(
  Ti = NULL,
  Si = NULL,
  Sr = NULL,
  ternary = FALSE,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)

LuEuSr(
  Lu = NULL,
  Eu = NULL,
  Sr = NULL,
  ternary = FALSE,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)

TiVSc(
  Ti = NULL,
  V = NULL,
  Sc = NULL,
  ternary = FALSE,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)

NbNaSr(
  Nb = NULL,
  Na = NULL,
  Sr = NULL,
  ternary = FALSE,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)

TiSmV(
  Ti = NULL,
  Sm = NULL,
  V = NULL,
  ternary = FALSE,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)

TiZrYSr(
  Ti = NULL,
  Zr = NULL,
  Y = NULL,
  Sr = NULL,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)

Pearce1976(
  major = NULL,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{Ti}{vector of Ti-concentrations (in ppm)}

\item{Si}{vector of Si-concentrations (in ppm)}

\item{Sr}{vector of Sr-concentrations (in ppm)}

\item{ternary}{logical. If \code{TRUE}, produces a ternary
diagram. Otherwise shows the results on a logratio plot.}

\item{pch}{plot character. See \code{?par} for details.}

\item{bg}{fill colour for the plot character.}

\item{show.labels}{logical. If \code{TRUE}, labels the polygonal
decision fields with text strings.}

\item{short}{use short labels when using the additional argument
\code{show.labels=TRUE}.}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}

\item{Lu}{vector of Lu-concentrations (in ppm)}

\item{Eu}{vector of Eu-concentrations (in ppm)}

\item{V}{vector of V-concentrations (in ppm)}

\item{Sc}{vector of Sc-concentrations (in ppm)}

\item{Nb}{vector of Nb-concentrations (in ppm)}

\item{Na}{vector of Na-concentrations (in ppm)}

\item{Sm}{vector of Sm-concentrations (in ppm)}

\item{Zr}{vector of Zr-concentrations (in ppm)}

\item{Y}{vector of Y-concentrations (in ppm)}

\item{major}{a data frame with the following columns: \code{SiO2},
\code{Al2O3}, \code{TiO2}, \code{CaO}, \code{MgO}, \code{MnO},
\code{K2O}, \code{Na2O} (in wt\%)}
}
\value{
a vector with tectonic affinities
}
\description{
discriminant analysis for oceanic basalts
}
\examples{
data(test,package='GeoplotR')
TiSiSr(Ti=wtpct2ppm(test$TiO2),Si=wtpct2ppm(test$SiO2),Sr=test$Sr)
}
\references{
Vermeesch, 2006. Tectonic discrimination diagrams
    revisited: Geochemistry, Geophysics, and Geosystems, 7, Q06017,
    doi:10.1029/2005GC001092
}
