% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{wtpct2ppm}
\alias{wtpct2ppm}
\title{wtpct2ppm}
\usage{
wtpct2ppm(x, oxide)
}
\arguments{
\item{x}{percentage by weight of the oxide}

\item{oxide}{Oxide of the element; one of 'SiO2', 'TiO2', 'Al2O3',
'Fe2O3', 'FeO', 'CaO', 'MgO', 'MnO', 'K2O', 'Na2O', 'P2O5'}
}
\value{
parts per million of the element by weight
}
\description{
Converts from percentage oxide by weight to parts per
million of the element
}
\examples{
ppm <- wtpct2ppm(3.08, 'TiO2')
}
