% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{xytest}
\alias{xytest}
\title{xy testing function}
\usage{
xytest(
  fname,
  xlim = NA,
  ylim = NA,
  log = "",
  short = FALSE,
  polygons = FALSE,
  ...
)
}
\arguments{
\item{fname}{path to a json file}

\item{xlim}{(optional) 2-element vector with the x-axis limits}

\item{ylim}{(optional) 2-element vetor with the y-axis limits}

\item{log}{one of

\code{''} linear axes

\code{'x'} semilogarithmic plot with linear y axis and logarithmic x axis

\code{'y'} semilogarithmic plot with linear x axis and logarithmic y axis

\code{'xy'} log-log plot}

\item{short}{logical. If \code{TRUE}, uses short labels}

\item{polygons}{logical. If \code{TRUE} plots the decision
polygons. If \code{TRUE}, plots the decision lines}

\item{...}{additional arguments to the generic plot function}
}
\value{
a list of text strings
}
\description{
tests json files with bivariate decision boundaries
}
\examples{
fn <- system.file('TiZrY.json',package='GeoplotR')
xyztest(fn)
}
