% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{ZrTi}
\alias{ZrTi}
\title{Ti-Zr}
\usage{
ZrTi(
  Zr = NULL,
  Ti = NULL,
  xlim = NULL,
  ylim = NULL,
  show.labels = TRUE,
  short = FALSE,
  ...
)
}
\arguments{
\item{Zr}{vector with Zr concentrations (ppm)}

\item{Ti}{vector with Ti concentrations (ppm)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{show.labels}{logical. If \code{TRUE}, labels the polygonal
decision fields with text strings.}

\item{short}{use short labels when using the additional argument
\code{show.labels=TRUE}.}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}
}
\value{
a vector with tectonic affinities
}
\description{
Zr vs Ti discrimination diagram of Dilek and Furnes
    (2009) for MORB, IAT, and boninites, modified after earlier
    workers.
}
\examples{
data(test,package='GeoplotR')
ZrTi(Zr=c(50,40,60),Ti=c(1000,4000,6000))
}
\references{
Dilek, Y. and Furnes, H., 2009, Structure and
    geochemistry of Tethyan ophiolites and their petrogenesis in
    subduction rollback systems.  Lithos, v. 113, p. 1-20.
}
