% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.R
\name{AFM}
\alias{AFM}
\title{A-F-M}
\usage{
AFM(
  A,
  F,
  M,
  ternary = TRUE,
  radial = FALSE,
  bty = "n",
  asp = 1,
  bw = "nrd0",
  decision = TRUE,
  dlty = 2,
  dlwd = 1.5,
  dcol = "blue",
  ...
)
}
\arguments{
\item{A}{vector with (Na\eqn{_2}O+K\eqn{_2}O) concentrations (in
wt\%)}

\item{F}{vector with (FeO + FeO\eqn{_2}O\eqn{_3}) concentrations
(in wt\%)}

\item{M}{vector with MgO concentrations (in wt\%)}

\item{ternary}{logical. If \code{FALSE}, produces a logratio plot.}

\item{radial}{logical. If \code{TRUE}, adds a kernel density
estimate of the Bowen-Fenner indices on a radial scale. Not
used if \code{ternary=FALSE}.}

\item{bty}{A character string that determines the type of `box'
which is drawn about plots. See \code{par}.}

\item{asp}{the y/x aspect ratio, see \code{plot.window}.}

\item{bw}{the smoothing bandwidth to be used for the kernel density
estimate. See \code{density}.}

\item{decision}{logical. If \code{TRUE}, adds Vermeesch and Pease's
two-stage decision boundary between the calc-alkaline and
tholeiitic magma series to the plot.}

\item{dlty}{line type of the decision boundary.}

\item{dlwd}{line width of the decision boundary.}

\item{dcol}{colour of the decision boundary.}

\item{...}{additional arguments for the generic \code{points}
function.}
}
\value{
the Bowen-Fenner indices of the samples, where positive
    values indicate calc-alkaline and negative numbers tholeiitic
    compositions (Vermeesch and Pease, 2021).
}
\description{
A-F-M diagram
}
\examples{
data(cath,package='GeoplotR')
Cascades <- (cath$affinity=='ca')
AFM(A=(cath$Na2O+cath$K2O)[Cascades],
    F=cath$FeOT[Cascades],M=cath$MgO[Cascades])
}
\references{
Irvine, T. N. and Baragar, W. A guide to the chemical
    classification of the common volcanic rocks. Canadian Journal
    of Earth Sciences, 8(5):523--548, 1971.

Kuno, H. Differentiation of basalt magmas. Basalts: The Poldervaart
    treatise on rocks of basaltic composition, pages 623--688, 1968.

Vermeesch, P. and Pease, V. A genetic classification of the
tholeiitic and calc-alkaline magma series, Geochemical Perspective
Letters (in review).
}
