% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.R
\name{AFM}
\alias{AFM}
\title{A-F-M}
\usage{
AFM(
  A,
  F,
  M,
  genetic = TRUE,
  ternary = TRUE,
  twostage = TRUE,
  kde = TRUE,
  decision = TRUE,
  bty = "n",
  asp = 1,
  xpd = FALSE,
  bw = "nrd0",
  pch = 21,
  bg = NULL,
  dlwd = 1.5,
  dcol = "blue",
  padding = 0.15,
  xlim = NULL,
  ylim = NULL,
  show.labels = FALSE,
  short = TRUE,
  ...
)
}
\arguments{
\item{A}{vector with (Na\eqn{_2}O+K\eqn{_2}O) concentrations (in
wt\%)}

\item{F}{vector with (FeO + FeO\eqn{_2}O\eqn{_3}) concentrations
(in wt\%)}

\item{M}{vector with MgO concentrations (in wt\%)}

\item{genetic}{logical. If \code{TRUE}, uses the logratio fit of
Vermeesch and Pease (2021). Otherwise uses the empirical fit of
Irvine and Baragar (1971).}

\item{ternary}{logical. If \code{FALSE}, produces a logratio plot.}

\item{twostage}{logical. If \code{TRUE}, applies the two-stage
magma evolution model of Vermeesch and Pease (2021). Otherwise
applies the single stage model. Only used if
\code{genetic=TRUE}.}

\item{kde}{logical. If \code{TRUE}, adds a kernel density estimate
of the Bowen-Fenner indices on a radial scale. Not used if
\code{ternary=FALSE}.}

\item{decision}{logical. If \code{TRUE}, adds the decision boundary
between the tholeiitic and calc-alkaline magma series to the
plot.}

\item{bty}{A character string that determines the type of `box'
which is drawn about plots. See \code{par}.}

\item{asp}{the y/x aspect ratio, see \code{plot.window}.}

\item{xpd}{clip the plot region? See \code{par} for further
details.}

\item{bw}{the smoothing bandwidth to be used for the kernel density
estimate. See \code{density}.}

\item{pch}{plotting character. See the documentation of
\code{points()} for details.}

\item{bg}{background colour for the plot symbols. If \code{NULL},
uses the assigned classes.}

\item{dlwd}{line width of the decision boundary.}

\item{dcol}{colour of the decision boundary.}

\item{padding}{fractional measure of distance between the data and
the kernel density estimate.}

\item{xlim}{the x axis limits of the plot (see \code{plot.default}
for details).}

\item{ylim}{the y axis limits of the plot (see \code{plot.default}
for details).}

\item{show.labels}{logical. If \code{TRUE}, labels the tholeiitic
and calc-alkaline fields.}

\item{short}{logical. If \code{TRUE}, uses abbreviated labels (only
used if \code{show.labels} is \code{TRUE}).}

\item{...}{additional arguments for the generic \code{points}
function.}
}
\value{
the Bowen-Fenner indices of the samples, where positive
    values indicate calc-alkaline and negative numbers tholeiitic
    compositions (Vermeesch and Pease, 2021).
}
\description{
A-F-M diagram
}
\examples{
data(cath,package='GeoplotR')

A <- cath[,'Na2O']+cath[,'K2O']
F <- cath[,'FeOT']
M <- cath[,'MgO']

# 1. Irvine and Baragar (1971):

AFM(A=A,F=F,M=M,genetic=FALSE,
    bg=cath[,'affinity'],dcol='blue',pch=21)

# 2. Vermeesch and Pease (2021):

oldpar <- par(mfrow=c(2,2),mar=c(4,4,0,0))

tern <- c(TRUE,FALSE,TRUE,FALSE)
stag <- c(TRUE,TRUE,FALSE,FALSE)
for (i in 1:4){
   AFM(A,F,M,ternary=tern[i],decision=TRUE,twostage=stag[i],
       pch=21,bw=0.2,padding=0.15,bg=cath[,'affinity'])
}

par(oldpar)

}
\references{
Irvine, T. N. and Baragar, W. A guide to the chemical
    classification of the common volcanic rocks. Canadian Journal
    of Earth Sciences, 8(5):523--548, 1971.

Kuno, H. Differentiation of basalt magmas. Basalts: The Poldervaart
    treatise on rocks of basaltic composition, pages 623--688, 1968.

Irvine, T.N. and Baragar, W.R.A., 1971. A guide to the chemical
classification of the common volcanic rocks. Canadian journal of
earth sciences, 8(5), pp.523-548.

Vermeesch, P. and Pease, V. A genetic classification of the
tholeiitic and calc-alkaline magma series, Geochemical Perspective
Letters.
}
