% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.R
\name{BF}
\alias{BF}
\title{BF}
\usage{
BF(A, F, M)
}
\arguments{
\item{A}{vector with (Na\eqn{_2}O+K\eqn{_2}O) concentrations (in
wt\%)}

\item{F}{vector with (FeO + FeO\eqn{_2}O\eqn{_3}) concentrations
(in wt\%)}

\item{M}{vector with MgO concentrations (in wt\%)}
}
\value{
the Bowen-Fenner indices of the samples, where positive
    values indicate calc-alkaline and negative numbers tholeiitic
    compositions (Vermeesch and Pease, 2021).
}
\description{
Bowen-Fenner index
}
\examples{
data(cath,package='GeoplotR')
bfi <- BF(cath$A,cath$F,cath$M)
oldpar <- par(mfrow=c(2,1))
on.exit(par(oldpar))
hist(bfi[cath$affinity=='ca'])
hist(bfi[cath$affinity=='th'])
}
\references{
Vermeesch, P. and Pease, V. A genetic classification of
    the tholeiitic and calc-alkaline magma series, Geochemical
    Perspective Letters (in review).
}
