% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{Pearce1984}
\alias{Pearce1984}
\alias{YNb}
\alias{YNbRb}
\alias{YbTa}
\alias{YbTaRb}
\title{Pearce et al. (1984)}
\usage{
YNb(Y = NULL, Nb = NULL, xlim = c(1, 1000), ylim = c(1, 1000), ...)

YNbRb(
  Y = NULL,
  Nb = NULL,
  Rb = NULL,
  xlim = c(1, 1000),
  ylim = c(1, 1000),
  ...
)

YbTa(Yb = NULL, Ta = NULL, xlim = c(0.1, 100), ylim = c(0.1, 100), ...)

YbTaRb(
  Yb = NULL,
  Ta = NULL,
  Rb = NULL,
  xlim = c(0.9, 110),
  ylim = c(1, 1000),
  ...
)
}
\arguments{
\item{Y}{vector with Y concentrations (ppm)}

\item{Nb}{vector with Nb concentrations (ppm)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}

\item{Rb}{vector with Rb concentrations (ppm)}

\item{Yb}{vector with Yb concentrations (ppm)}

\item{Ta}{vector with Ta concentrations (ppm)}
}
\value{
a vector with tectonic affinities
}
\description{
tectonic discrimination diagram for granites
}
\examples{
data(test,package='GeoplotR')
YNb(Y=test[,'Y'],Nb=test[,'Nb'])
YNbRb(Y=test[,'Y'],Nb=test[,'Nb'],Rb=test[,'Rb'])
YbTa(Yb=test[,'Yb'],Ta=test[,'Ta'])
YbTaRb(Yb=test[,'Yb'],Ta=test[,'Ta'],Rb=test[,'Rb'])
}
\references{
Pearce, J. A., Harris, N. B. W., and Tindle, A. G.,
    1984,
    "Trace Element Discrimination Diagrams for the Tectonic Interpretation of Granitic Rocks":
    Journal of Petrology, v. 25, no. 4, p. 956-983.
}
