% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{TAS}
\alias{TAS}
\title{TAS diagram}
\usage{
TAS(
  Na2O,
  K2O,
  SiO2,
  plot = TRUE,
  xlim = c(35, 90),
  ylim = c(0, 20),
  labels = FALSE,
  volcanic = TRUE,
  ...
)
}
\arguments{
\item{Na2O}{vector with Na\eqn{_2}O concentrations (wt\%)}

\item{K2O}{vector with K\eqn{_2}O concentrations (wt\%)}

\item{SiO2}{vector with SiO\eqn{_2} concentrations (wt\%)}

\item{plot}{logical. If \code{FALSE}, suppresses the graphical
output.}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{labels}{logical. Label to mark the different fields of the
TAS diagram.}

\item{volcanic}{logical. Set to \code{FALSE} for plutonic rocks.}

\item{...}{additional arguments to the generic \code{points}
function.}
}
\value{
a vector with rock types
}
\description{
Total Alkali Silica diagram
}
\examples{
data(test,package='GeoplotR')
TAS(test[,'Na2O'],test[,'K2O'],test[,'SiO2'])
}
