% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{TAS}
\alias{TAS}
\title{TAS diagram}
\usage{
TAS(
  Na2O = NULL,
  K2O = NULL,
  SiO2 = NULL,
  xlim = c(35, 80),
  ylim = c(0, 15),
  volcanic = TRUE,
  ...
)
}
\arguments{
\item{Na2O}{vector with Na\eqn{_2}O concentrations (wt\%)}

\item{K2O}{vector with K\eqn{_2}O concentrations (wt\%)}

\item{SiO2}{vector with SiO\eqn{_2} concentrations (wt\%)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{volcanic}{logical. Set to \code{FALSE} for plutonic rocks.}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}
}
\value{
a vector with rock types
}
\description{
Total Alkali Silica diagram
}
\examples{
data(test,package='GeoplotR')
TAS(test[,'Na2O'],test[,'K2O'],test[,'SiO2'])
}
\references{
Le Maitre, R. W., Streckeisen, A., Zanettin, B.,
Bas, M. J. L., Bonin, B., and Bateman, P., 2004,
"Igneous Rocks: A Classification and Glossary of Terms":
Cambridge University Press, v. -1, no. 70, p. 93–120.
}
