% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{TAS}
\alias{TAS}
\title{TAS diagram}
\usage{
TAS(
  Na2O = NULL,
  K2O = NULL,
  SiO2 = NULL,
  xlim = NULL,
  ylim = NULL,
  plutonic = FALSE,
  show.labels = TRUE,
  short = TRUE,
  ...
)
}
\arguments{
\item{Na2O}{vector with Na\eqn{_2}O concentrations (wt\%)}

\item{K2O}{vector with K\eqn{_2}O concentrations (wt\%)}

\item{SiO2}{vector with SiO\eqn{_2} concentrations (wt\%)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{plutonic}{logical. If \code{TRUE}, uses the nomenclature of
Middlemost (1994).}

\item{show.labels}{logical. If \code{TRUE}, labels the
discrimination fields on the plot.}

\item{short}{use short labels when using the additional argument
\code{show.labels=TRUE}.}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}
}
\value{
a vector with rock types
}
\description{
Total Alkali Silica diagram
}
\examples{
data(test,package='GeoplotR')
TAS(test[,'Na2O'],test[,'K2O'],test[,'SiO2'])
}
\references{
Le Maitre, R. W., Streckeisen, A., Zanettin, B., Bas,
    M. J. L., Bonin, B., and Bateman, P., 2004,
    "Igneous Rocks: A Classification and Glossary of Terms":
    Cambridge University Press, v. -1, no. 70, p. 93–120.

Middlemost, E.A., 1994. Naming materials in the magma/igneous rock
system. Earth-science reviews, 37(3-4), pp.215-224.
}
