% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tectodisc.R
\name{TiZrY}
\alias{TiZrY}
\title{Ti-Zr-Y}
\usage{
TiZrY(
  Ti = NULL,
  Zr = NULL,
  Y = NULL,
  unit = c("wt\%", "ppm", "ppm"),
  type = c("LDA", "QDA", "Pearce"),
  plot = c("none", "ternary", "logratio"),
  ...
)
}
\arguments{
\item{Ti}{vector with Ti (ppm) or TiO2 (wt\%) concentrations}

\item{Zr}{vector with Zr concentrations (ppm)}

\item{Y}{vector with Y concentrations (ppm)}

\item{unit}{3-element vector indicating whether the Ti, Zr and Y
concentrations are supplied as ppm or weight percent.}

\item{type}{either \code{'LDA'} for linear discriminant analysis,
\code{'QDA'} for quadratic discriminant analysis, or
\code{'Pearce'} for the nominal decision boundaries of Pearce
and Cann (1973). The latter option has not been implemented
yet.}

\item{plot}{either \code{'none'} to omit the plot, \code{'ternary'}
for a ternary diagram, or \code{'logratio'} for a bivariate
logratio plot}

\item{...}{additional arguments for the generic \code{points}
function.}
}
\value{
if \code{type='LDA'} or \code{type='QDA'}, a list with
    components \code{class}, \code{posterior} and \code{x};
    otherwise a table with labels for \code{MORB}, \code{IAB} and
    \code{OIB}.
}
\description{
Ti-Zr-Y tectonic discrimination diagram
}
\examples{
data(test,package='GeoplotR')
TiZrY(Ti=test[,'TiO2'],Zr=test[,'Zr'],Y=test[,'Y'],
      unit=c('wt\%','ppm','ppm'),
      type='QDA',plot='ternary')
}
\references{
Pearce, J. A., and Cann, J. R., 1973, Tectonic setting
    of basic volcanic rocks determined using trace element
    analyses: Earth and Planetary Science Letters, v. 19, no. 2,
    p. 290-300.
}
