% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{xyztest}
\alias{xyztest}
\title{xyz testing function}
\usage{
xyztest(fname, short = FALSE, polygons = FALSE, ...)
}
\arguments{
\item{fname}{path to a json file}

\item{short}{logical. If \code{TRUE}, uses short labels}

\item{polygons}{logical. If \code{TRUE} plots the decision
polygons. If \code{TRUE}, plots the decision lines}

\item{...}{additional arguments to the generic plot function}
}
\value{
a list of text strings
}
\description{
tests json files with ternary decision boundaries
}
\examples{
fn <- system.file('TiZrY.json',package='GeoplotR')
xyztest(fn)
}
