% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vermeesch2006.R
\name{ZrTi}
\alias{ZrTi}
\title{Ti-Zr}
\usage{
ZrTi(
  Zr = NULL,
  Ti = NULL,
  type = c("LDA", "QDA", "Pearce", "Dilek"),
  ternary = FALSE,
  pch = 21,
  bg = NULL,
  show.labels = FALSE,
  short = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{Zr}{vector with Zr concentrations (ppm)}

\item{Ti}{vector with Ti concentrations (ppm)}

\item{type}{either \code{'LDA'} for linear discriminant analysis,
\code{'QDA'} for quadratic discriminant analysis,
\code{'Pearce'} for the nominal decision boundaries of Pearce
and Cann (1973), or \code{'Dilek'} for the nominal decision
boundaries of Dilek and Furnes (2009).}

\item{ternary}{logical. If \code{TRUE}, produces a ternary diagram.
Only used if \code{type} is \code{LDA} or \code{QDA}.}

\item{pch}{plot character. See \code{?par} for details.}

\item{bg}{fill colour for the plot character.}

\item{show.labels}{logical. If \code{TRUE}, labels the polygonal
decision fields with text strings.}

\item{short}{use short labels when using the additional argument
\code{show.labels=TRUE}.}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{...}{additional arguments to the generic \code{points}
function, may include the logical argument \code{show.labels}
which labels the fields in the diagram.}
}
\value{
a vector with tectonic affinities
}
\description{
Ti-Zr tectonic discrimination diagram.
}
\examples{
data(test,package='GeoplotR')
ZrTi(Zr=c(50,40,60),Ti=c(1000,4000,6000))
}
\references{
Pearce, J.A. and Cann, J.R., 1973. Tectonic setting of
    basic volcanic rocks determined using trace element
    analyses. Earth and Planetary Science Letters, 19(2),
    pp. 290-300.

Dilek, Y. and Furnes, H., 2009, Structure and geochemistry of
Tethyan ophiolites and their petrogenesis in subduction rollback
systems.  Lithos, v. 113, p. 1-20.
}
